+++ Sources/KeyboardManager/KeyboardManager.swift
@@ -71,6 +71,9 @@ open class KeyboardManager: NSObject, UIGestureRecognizerDelegate {
     /// A cached notification used as a starting point when a user dragging the `scrollView` down
     /// to interactively dismiss the keyboard
     private var cachedNotification: KeyboardNotification?
+    
+    /// Used to fix a glitch that would otherwise occur when using pagesheets on iPad in iOS 14
+    private var justDidWillHide = false
 
     // MARK: - Initialization
 
@@ -139,6 +142,15 @@ open class KeyboardManager: NSObject, UIGestureRecognizerDelegate {
         callbacks[event] = callback
         return self
     }
+    
+    /// When e.g. using pagesheets on iPad the inputAccessoryView is not stuck to the bottom of the screen.
+    /// This value represents the size of the gap between the bottom of the screen and the bottom of the inputAccessoryView.
+    private var bottomGap: CGFloat {
+        if let inputAccessoryView = inputAccessoryView, let window = inputAccessoryView.window, let superView = inputAccessoryView.superview {
+            return window.frame.height - window.convert(superView.frame, to: window).maxY
+        }
+        return 0
+    }
 
     /// Constrains the `inputAccessoryView` to the bottom of its superview and sets the
     /// `.willChangeFrame` and `.willHide` event callbacks such that it mimics an `InputAccessoryView`
@@ -169,9 +181,21 @@ open class KeyboardManager: NSObject, UIGestureRecognizerDelegate {
             else { return }
 
             let keyboardHeight = notification.endFrame.height
-            self?.animateAlongside(notification) {
-                self?.constraints?.bottom?.constant = -keyboardHeight - (additionalBottomSpace?() ?? 0)
-                self?.inputAccessoryView?.superview?.layoutIfNeeded()
+            let animateAlongside = {
+                self?.animateAlongside(notification) {
+                    self?.constraints?.bottom?.constant = min(0, -keyboardHeight + (self?.bottomGap ?? 0)) - (additionalBottomSpace?() ?? 0)
+                    self?.inputAccessoryView?.superview?.layoutIfNeeded()
+                }
+            }
+            animateAlongside()
+
+            // Trigger a new animation if gap changed, this typically happens when using pagesheet on portrait iPad
+            let initialBottomGap = self?.bottomGap ?? 0
+            DispatchQueue.main.async {
+                let newBottomGap = self?.bottomGap ?? 0
+                if newBottomGap != 0 && newBottomGap != initialBottomGap {
+                    animateAlongside()
+                }
             }
         }
         callbacks[.willChangeFrame] = { [weak self] (notification) in
@@ -179,17 +203,35 @@ open class KeyboardManager: NSObject, UIGestureRecognizerDelegate {
             guard
                 self?.isKeyboardHidden == false,
                 notification.isForCurrentApp
-            else { return }
-
-            self?.animateAlongside(notification) {
-                self?.constraints?.bottom?.constant = -keyboardHeight - (additionalBottomSpace?() ?? 0)
-                self?.inputAccessoryView?.superview?.layoutIfNeeded()
+            else {
+                return
+            }
+            let animateAlongside = {
+                self?.animateAlongside(notification) {
+                    self?.constraints?.bottom?.constant = min(0, -keyboardHeight + (self?.bottomGap ?? 0)) - (additionalBottomSpace?() ?? 0)
+                    self?.inputAccessoryView?.superview?.layoutIfNeeded()
+                }
+            }
+            animateAlongside()
+            
+            // Trigger a new animation if gap changed, this typically happens when using pagesheet on portrait iPad
+            let initialBottomGap = self?.bottomGap ?? 0
+            DispatchQueue.main.async {
+                let newBottomGap = self?.bottomGap ?? 0
+                if newBottomGap != 0 && newBottomGap != initialBottomGap && !(self?.justDidWillHide ?? false) {
+                    animateAlongside()
+                }
             }
         }
         callbacks[.willHide] = { [weak self] (notification) in
+            guard notification.isForCurrentApp else { return }
+            self?.justDidWillHide = true
             self?.animateAlongside(notification) { [weak self] in
                 self?.constraints?.bottom?.constant = self?.additionalInputViewBottomConstraintConstant() ?? 0
                 self?.inputAccessoryView?.superview?.layoutIfNeeded()
+            }
+            DispatchQueue.main.async {
+                self?.justDidWillHide = false
             }
         }
         return self
@@ -311,7 +353,7 @@ open class KeyboardManager: NSObject, UIGestureRecognizerDelegate {
         frame.size.height = window.bounds.height - frame.origin.y
         keyboardNotification.endFrame = frame
 
-        var yCoordinateDirectlyAboveKeyboard = -frame.height
+        var yCoordinateDirectlyAboveKeyboard = -frame.height + bottomGap
         if shouldApplyAdditionBottomSpaceToInteractiveDismissal, let additionalBottomSpace = additionalBottomSpace {
             yCoordinateDirectlyAboveKeyboard -= additionalBottomSpace()
         }